Refactored directories
